async.QObject: fix method signature for inner objs
QObject instances that are created because they are properties of
directly exposed objects should behave the same as the directly exposed
objects: their methods should also return a future.

For simplicity, patch the PlainQObject so that its _unwrapQObject()
method uses a sort of "virtual constructor": all child QObjects created
are the same type as the calling object, so async.QObject in the async
case and PlainQObject in the plain case.